backend/chatbot/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
python-dotenv
beautifulsoup4
markdown
faiss-cpu
langchain_community
langchain-experimental
